refactor: extract ChoiceGroup types into shared module and update imports across codebase
